aarch32-symbolic: Factor register-related definitions into a separate module
